requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyJWT==2.10.1
bcrypt==4.2.1
python-multipart==0.0.20
